+++ injitest/src/main/java/io/mosip/test/mob/inji/api/JsonPrecondtion.java
@@ -1,4 +1,4 @@
-package io.mosip.test.mob.inji.authentication.fw.precon;
+package io.mosip.test.mob.inji.api;
 
 import org.apache.commons.beanutils.PropertyUtils;
 
@@ -11,14 +11,14 @@ import com.google.gson.JsonParser;
 
 public class JsonPrecondtion {
 	private static final Logger JSONPRECONDATION_LOGGER = Logger.getLogger(JsonPrecondtion.class);
-	
+
 	public static String toPrettyFormat(String jsonString) {
 		Gson gson = new GsonBuilder().setPrettyPrinting().create();
 		JsonParser jp = new JsonParser();
 		JsonElement je = jp.parse(jsonString);
 		return gson.toJson(je);
 	}
-	
+
 	public static String getValueFromJson(String jsonContent, String fieldMapper) {
 		try {
 			ObjectMapper mapper = new ObjectMapper();
@@ -26,7 +26,7 @@ public class JsonPrecondtion {
 			return PropertyUtils.getProperty(jsonObj, fieldMapper).toString();
 		} catch (Exception expection) {
 			JSONPRECONDATION_LOGGER
-					.error("Exception Occured in retrieving the value from json file: " + expection.getMessage());
+			.error("Exception Occured in retrieving the value from json file: " + expection.getMessage());
 			return "Cannot retrieve data or content for the object mapper from  JSON";
 		}
 	}
