+++ injitest/src/main/java/io/mosip/test/mob/inji/api/pojo/From.java
@@ -1,8 +1,8 @@
-package io.mosip.test.mob.inji.utility.pojo;
+package io.mosip.test.mob.inji.api.pojo;
 
 import java.util.ArrayList;
 
-public class Cc{
+public class From{
 	 public ArrayList<Value> value;
 	 public String html;
 	 public String text;
