+++ injitest/src/main/java/io/mosip/test/mob/inji/api/pojo/Headers.java
@@ -1,0 +1,4 @@
+package io.mosip.test.mob.inji.api.pojo;
+
+public class Headers{
+}
