+++ injitest/src/main/java/io/mosip/test/mob/inji/api/pojo/To.java
@@ -1,8 +1,8 @@
-package io.mosip.test.mob.inji.utility.pojo;
+package io.mosip.test.mob.inji.api.pojo;
 
 import java.util.ArrayList;
 
-public class From{
+public class To{
 	 public ArrayList<Value> value;
 	 public String html;
 	 public String text;
