+++ injitest/src/main/java/io/mosip/test/mob/inji/testcases/VcDownloadAndVerifyUsingVidTest.java
@@ -2,6 +2,7 @@ package io.mosip.test.mob.inji.testcases;
 
 import org.testng.annotations.Test;
 
+import io.mosip.test.mob.inji.api.BaseTestCase;
 import io.mosip.test.mob.inji.driver.TestRunner;
 import io.mosip.test.mob.inji.pages.*;
 import io.mosip.test.mob.inji.utils.TestDataReader;
@@ -32,7 +33,7 @@ public class VcDownloadAndVerifyUsingVidTest extends BaseTest {
         RetrieveIdPage retrieveIdPage = homePage.downloadCard();
 
         assertTrue(retrieveIdPage.isRetrieveIdPageLoaded(), "Verify if retrieve id page is displayed");
-        OtpVerification otpVerification = retrieveIdPage.clickOnVid(target).setEnterIdTextBox(TestRunner.perpetualVid).clickOnGenerateCardButton();
+        OtpVerification otpVerification = retrieveIdPage.clickOnVid(target).setEnterIdTextBox(BaseTestCase.perpetualVid).clickOnGenerateCardButton();
         assertTrue(otpVerification.isOtpVerificationPageLoaded(), "Verify if otp verification page is displayed");
         otpVerification.enterOtp(NewOtp, target);
