--- injitest/src/main/java/io/mosip/test/mob/inji/utility/pojo/Headers.java
@@ -1,4 +1,0 @@
-package io.mosip.test.mob.inji.utility.pojo;
-
-public class Headers{
-}
