+++ injitest/src/main/resources/TestData.json
@@ -1,13 +1,15 @@
 {
+  "uin": "6187360582",
+  "vid": "5037415473849368",
   "externalemail": "Resident_AddIdentity_ValidParam_smoke_Pos@mosip.net",
+   "otp": "1234",
   "passcode": "111111",
   "fullName": "santush",
   "idType": "National card",
   "status": "Valid",
   "generatedOn": "9/5/2023",
   "aid": "10001100660000620230714065538",
-"setExcludedGroups":"AVT,PVT",
-  "androidAppPath":"D:/GITINJI/inji/injitest/apps/android/Inji_latest.apk",
-  "androidDevice": "Pixel_6a_API_30"
-  
+  "setExcludedGroups":"",
+  "androidAppPath":"D:/JarInji/JarInji/apps/android/Inji_latest.apk",
+  "androidDevice": "Pixel_3a_API_34_extension_level_7_x86_64"
 }
